Safari-611.1.10/LayoutTests/accessibility/set-value-not-work-for-disabled-sliders-expected.txt


This tests that disabled sliders can't be incremented or decremented.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS slider.intValue is 50
PASS slider.intValue is 50
PASS slider.intValue is 50
PASS slider.intValue is 50
PASS slider.intValue is 50
PASS slider.intValue is 50
PASS slider.intValue is 50
PASS slider.intValue is 50
PASS slider.intValue is 50
PASS slider.intValue is 50
PASS slider.intValue is 55
PASS slider.intValue is 45
PASS slider.intValue is 50
PASS slider.intValue is 55
PASS slider.intValue is 45
PASS slider.intValue is 50
PASS slider.intValue is 55
PASS slider.intValue is 45
PASS successfullyParsed is true

TEST COMPLETE
